Update kernel to last version

+++ kernel/kernel_device_modules-6.6/drivers/tee/tzdev/6/tz_common.h
@@ -355,7 +355,7 @@ enum cluster_type {
 struct sysconf_cluster_info {
 	uint32_t type;		/**< Cluster type (see enum cluster_type) */
 	uint32_t nr_cpus;	/**< Number of cpus in cluster */
-	uint32_t mask;		/**< Available CPUs in cluster */
+	uint64_t mask;		/**< Available CPUs in cluster */
 } __attribute__((__packed__));
 
 /**
